Redesigned News menu item

+++ adminlte/views.py
@@ -1,10 +1,11 @@
 from django.db import transaction
-from django.http import HttpResponse
+from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render, redirect, get_object_or_404
+from django.template.loader import render_to_string
 from django.views.decorators.http import require_POST
 
 from .utils.translation import get_translation
-from adminlte.models import Publication, Images, PublicationType
+from adminlte.models import Publication, Images, PublicationType, FiledValueTranslation
 from .forms import PublicationForm, SeoMetadataForm
 
 menu = [
@@ -13,7 +14,7 @@ menu = [
     {'title': 'Фильмы', 'url_name': 'movies', 'group': ['movies'], 'icon': 'fa-film'},
     {'title': 'Кинотеатры', 'url_name': 'cinemas', 'group': ['cinemas'], 'icon': 'fa-building'},
     {'title': 'Новости', 'url_name': 'news', 'group': ['news', 'add_news', 'edit_news'], 'icon': 'fa-newspaper'},
-    {'title': 'Акции', 'url_name': 'shares', 'group': ['shares','add_shares','edit_shares'], 'icon': 'fa-tags'},
+    {'title': 'Акции', 'url_name': 'shares', 'group': ['shares', ], 'icon': 'fa-tags'},  # 'add_shares','edit_shares'
     {'title': 'Страницы', 'url_name': 'pages', 'group': ['pages'], 'icon': 'fa-file-alt'},
     {'title': 'Пользователи', 'url_name': 'users', 'group': ['users'], 'icon': 'fa-users'},
     {'title': 'Рассылка', 'url_name': 'mailing', 'group': ['mailing'], 'icon': 'fa-envelope'},
@@ -24,8 +25,15 @@ def statistics(request):
     return render(request, "adminlte/pages/statistics.html", context={'menu': menu})
 
 
+
+
 def banners_sliders(request):
-    return render(request, 'adminlte/pages/banners_sliders_table.html', context={'menu': menu})
+    return render(request, 'adminlte/pages/banners_sliders_display.html', context={'menu': menu})
+
+def banners_sliders_add(request):
+    return render(request, "adminlte/pages/add/banners_sliders_add.html", context={'menu': menu})
+
+
 
 
 def movies(request):
@@ -46,20 +54,8 @@ def news_add(request):
         request=request,
         publication_type=PublicationType.NEWS,
         template_name='adminlte/pages/add/base_add.html',
-        translation_key=PublicationType.NEWS,
         redirect_url='news'
     )
-
-
-@require_POST
-def delete_news(request, pk):
-    return delete_publication(
-        request=request,
-        pk=pk,
-        publication_type=PublicationType.NEWS,
-        redirect_url='news'
-    )
-
 
 def edit_news(request, pk):
     return edit_publication(
@@ -67,44 +63,19 @@ def edit_news(request, pk):
         pk=pk,
         publication_type=PublicationType.NEWS,
         template_name='adminlte/pages/edit/base_edit.html',
-        translation_key=PublicationType.NEWS,
         redirect_url='news'
     )
 
-
-def shares(request):
-    publications = Publication.objects.filter(publication_type=PublicationType.SHARES)
-    return render(request, 'adminlte/pages/shares_tabel.html', context={'menu': menu, 'publications': publications})
-
-def add_shares(request):
-    return add_publication(
-        request=request,
-        publication_type=PublicationType.SHARES,
-        template_name='adminlte/pages/add/base_add.html',
-        translation_key=PublicationType.SHARES,
-        redirect_url='shares'
-    )
-
-
-@require_POST
-def delete_shares(request, pk):
+def delete_news(request, pk):
     return delete_publication(
         request=request,
         pk=pk,
-        publication_type=PublicationType.SHARES,
-        redirect_url='shares'
     )
 
 
-def edit_shares(request, pk):
-    return edit_publication(
-        request=request,
-        pk=pk,
-        publication_type=PublicationType.SHARES,
-        template_name='adminlte/pages/edit/base_edit.html',
-        translation_key=PublicationType.SHARES,
-        redirect_url='shares'
-    )
+
+def shares(request):
+    return render(request, 'adminlte/pages/shares_tabel.html', context={'menu': menu})
 
 
 def pages(request):
@@ -118,14 +89,13 @@ def users(request):
 def mailing(request):
     return render(request, 'adminlte/pages/mailing.html', context={'menu': menu})
 
-
-def add_publication(request, publication_type, template_name, translation_key, redirect_url):
+def add_publication(request, publication_type, template_name, redirect_url):
     language = request.GET.get('lang', 'ru')
-    t = get_translation(translation_key, language)
 
     if request.method == "POST":
-        form = PublicationForm(request.POST, request.FILES, prefix='pub')
-        seo_form = SeoMetadataForm(request.POST, prefix='seo')
+        form = PublicationForm(request.POST, request.FILES, prefix="pub")
+        seo_form = SeoMetadataForm(request.POST, prefix="seo")
+
         if form.is_valid() and seo_form.is_valid():
             with transaction.atomic():
                 seo_instance = seo_form.save()
@@ -135,30 +105,29 @@ def add_publication(request, publication_type, template_name, translation_key, r
                 publication.save()
                 form.save_m2m()
 
-                for f in request.FILES.getlist('gallery_images'):
+                for f in request.FILES.getlist("gallery_images"):
                     img = Images.objects.create(image_url=f)
                     publication.gallery.add(img)
 
             return redirect(redirect_url)
     else:
-        form = PublicationForm(prefix='pub')
-        seo_form = SeoMetadataForm(prefix='seo')
+        form = PublicationForm(prefix="pub")
+        seo_form = SeoMetadataForm(prefix="seo")
 
     return render(request, template_name, {
-        'menu': menu,
-        'form': form,
-        'seo_form': seo_form,
-        't': t,
-        'lang': language,
+        "form": form,
+        "seo_form": seo_form,
+        "lang": language,
+        "publication_type_label": publication_type.label,
+        "menu": menu,
     })
 
 
-def edit_publication(request, pk, publication_type, template_name, translation_key, redirect_url):
+def edit_publication(request, pk, publication_type, template_name, redirect_url):
     publication = get_object_or_404(Publication, pk=pk)
     seo_instance = publication.seo
 
     language = request.GET.get('lang', 'ru')
-    t = get_translation(translation_key, language)
 
     if request.method == "POST":
         form = PublicationForm(request.POST, request.FILES, instance=publication, prefix='pub')
@@ -169,9 +138,11 @@ def edit_publication(request, pk, publication_type, template_name, translation_k
                 seo = seo_form.save()
                 pub = form.save(commit=False)
                 pub.seo = seo
+
+                pub.published_at = form.cleaned_data.get('published_at')
+
                 pub.save()
                 form.save_m2m()
-
 
                 images_to_delete = request.POST.get("images_to_delete", "")
                 if images_to_delete:
@@ -179,7 +150,6 @@ def edit_publication(request, pk, publication_type, template_name, translation_k
                     for image in Images.objects.filter(id__in=ids):
                         pub.gallery.remove(image)
                         image.delete()
-
 
                 for f in request.FILES.getlist('gallery_images'):
                     img = Images.objects.create(image_url=f)
@@ -192,26 +162,30 @@ def edit_publication(request, pk, publication_type, template_name, translation_k
         seo_form = SeoMetadataForm(instance=seo_instance, prefix='seo')
 
     return render(request, template_name, {
-        'menu': menu,
         'form': form,
         'seo_form': seo_form,
-        't': t,
+        'menu': menu,
         'lang': language,
         'gallery_images': publication.gallery.all(),
+        'publication_type_label': publication_type.label,
     })
 
+def delete_publication(request, pk):
+    if request.method != 'POST':
+        return JsonResponse({'success': False, 'error': 'Invalid request method'}, status=405)
 
-@require_POST
-def delete_publication(request, pk, publication_type, redirect_url):
     publication = get_object_or_404(Publication, pk=pk)
     seo = publication.seo
     images = list(publication.gallery.all())
 
-    publication.delete()
-    if seo:
-        seo.delete()
-    for img in images:
-        img.delete()
+    try:
+        publication.delete()
+        if seo:
+            seo.delete()
+        for img in images:
+            img.delete()
 
-    return redirect(redirect_url)
+        return JsonResponse({"success": True})
+    except Exception as e:
+        return JsonResponse({"success": False, "error": str(e)}, status=500)
